% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.utils-package.R
\docType{package}
\name{pca.utils-package}
\alias{pca.utils}
\alias{pca.utils-package}
\title{Utilities for plotting PCA results and projecting in an existing PCA space.}
\description{
The \pkg{pca.utils} package provides two main functions
\itemize{
   \item \code{\link{plot_pca}}
   \item \code{\link{project_pca}}
}
}

