% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA results}
\usage{
plot_pca(x, npcs = c(1, 2), xlab = NULL, ylab = NULL, zlab = NULL,
  main = "PCA results", pch = NULL, col = NULL, bg = NULL, cex = NULL,
  add.legend = FALSE, legend.position = "bottomleft", legend.text = NULL,
  legend.col = NULL, legend.bg = NULL)
}
\arguments{
\item{x:}{data matrix or prcomp object to plot.}

\item{npcs:}{numeric vector indicating PCs to plot, defaults to c(1,2). If length(npcs)>3, only the first 3 elements will be used.}

\item{xlab:}{xlabel, see \link[graphics]{par}.}

\item{ylab:}{ylabel, see \link[graphics]{par}, see \link[graphics]{par}.}

\item{zlab:}{zlabel, see \link[rgl]{plot3d}.}

\item{main:}{title for the plot, see \link[graphics]{par}. Defaults to 'PCA results'.}

\item{pch:}{numeric vector of plotting characters, see \link[graphics]{par}.}

\item{col:}{vector of plotting character colors, see \link[graphics]{par}.}

\item{bg:}{vector of plotting character background colors, see \link[graphics]{par}.}

\item{cex:}{plotting character size. Defaults to 1 for 2D and to 7 for 3D plots.}

\item{add.legend:}{add a legend to the plot, defaults to FALSE.}

\item{legend.position:}{, argument to \link[graphics]{legend}. Position of the legend in the plot, defaults to 'bottomleft'.}

\item{legend.text:}{character vector to appear in the legend.}

\item{legend.col:}{color of legend elements, defaults to unique(col)}

\item{legend.bg:}{background color of legend elements, defaults to unique(bg)}
}
\value{
a list with class "prcomp", containing the same elements as \link[stats]{prcomp}:
\itemize{
 \item var_explained: numeric vector indicating the varince explained by each PC
}
}
\description{
inputs:
}
\examples{

data(iris)

# 2D PCA
plot_pca(x=iris[1:4],col=iris$Species)

# 3D PCA
plot_pca(x=iris[1:4],npcs=3,col=iris$Species)

}

