% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pca.R
\name{project_pca}
\alias{project_pca}
\title{Project new data in existing PCA space}
\usage{
project_pca(Xnew = NULL, pc = NULL)
}
\arguments{
\item{Xnew:}{new data matrix to project.}

\item{pc:}{prcomp object used for projection generated by \link[stats]{prcomp}. Must have the same number of columns as X.}
}
\value{
Xpc: a matrix that corresponds to the projection of X in the PCA space. Xpc[,1] is the first PC, Xpc[,2] the second PC, etc.
}
\description{
inputs:
}
\examples{

set.seed(1)
data(iris)

#randomly select 100 data points from the iris data set
#to perform PCA
ix=rep(FALSE,nrow(iris))
ix[sample(nrow(iris),100)]=TRUE

# perform PCA
pc=prcomp(as.matrix(iris[ix,1:4]))

# plot the PCA
plot_pca(pc,col=iris$Species[ix])

# project the remaining 50 data points (given by !ix) in the existing PCA space
Xpc=project_pca(as.matrix(iris[!ix,1:4]),pc)

#plot the new data points in the existing plot
points(Xpc[,1],Xpc[,2])
}

